refactor: split org-cache-service into auth and metadata modules
Move Clerk identity caching (slug, name) to auth/org-cache.ts and
business data functions (getOrgMetadata, getOrgBillingPeriod) to
zero/org/org-metadata-service.ts. Simplify org-cache-service.ts to
a thin composition layer. Remove org_cache caching layer from
getOrgBillingPeriod so it reads org_metadata directly.

Closes #7791
