+++ turbo/apps/web/src/lib/zero/billing/usage-service.ts
@@ -1,6 +1,6 @@
 import { sql, and, eq, gte, lt, inArray } from "drizzle-orm";
 import { type MemberUsage, type UsageMembersResponse } from "@vm0/core";
-import { getOrgBillingPeriod } from "../org/org-cache-service";
+import { getOrgBillingPeriod } from "../org/org-metadata-service";
 import { creditUsage } from "../../../db/schema/credit-usage";
 import { userCache } from "../../../db/schema/user-cache";
 import { orgMembersMetadata } from "../../../db/schema/org-members-metadata";
